+++ InstagramCloneSwiftUI/Core/Components/PostGridView.swift
@@ -1,0 +1,35 @@
+//
+//  PostGridView.swift
+//  InstagramCloneSwiftUI
+//
+//  Created by Alejandro Robles on 19/06/23.
+//
+
+import SwiftUI
+
+struct PostGridView: View {
+    var posts: [Post]
+    private let gridItems: [GridItem] = [
+        .init(.flexible(), spacing: 1),
+        .init(.flexible(), spacing: 1),
+        .init(.flexible(), spacing: 1)
+    ]
+    private let imageDimension: CGFloat = UIScreen.main.bounds.width/3 - 1
+    var body: some View {
+        LazyVGrid(columns: gridItems, spacing: 2) {
+            ForEach(posts) { post in
+                Image(post.imageURL)
+                    .resizable()
+                    .scaledToFill()
+                    .frame(width: imageDimension, height: imageDimension)
+                    .clipped()
+            }
+        }
+    }
+}
+
+struct PostGridView_Previews: PreviewProvider {
+    static var previews: some View {
+        PostGridView(posts: Post.MOCK_POSTS)
+    }
+}
