+++ InstagramCloneSwiftUI/Core/Profile/ProfileView.swift
@@ -9,11 +9,6 @@ import SwiftUI
 
 struct ProfileView: View {
     let user: User
-    private let gridItems: [GridItem] = [
-        .init(.flexible(), spacing: 1),
-        .init(.flexible(), spacing: 1),
-        .init(.flexible(), spacing: 1)
-    ]
     var posts: [Post] {
         return Post.MOCK_POSTS.filter { $0.user?.userName == user.userName }
     }
@@ -62,13 +57,7 @@ struct ProfileView: View {
                 }
                 Divider()
             }
-            LazyVGrid(columns: gridItems, spacing: 2) {
-                ForEach(posts) { post in
-                    Image(post.imageURL)
-                        .resizable()
-                        .scaledToFill()
-                }
-            }
+            PostGridView(posts: posts)
         }
         .navigationTitle("Profile")
         .navigationBarTitleDisplayMode(.inline)
